pyproject.toml
[project]
name = "data-visor"
version = "0.1.0"
description = "DataVisor - Unified CV dataset introspection tool"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "duckdb>=1.4.4",
    "einops>=0.8.0",
    "fastapi>=0.128.7",
    "fsspec>=2026.2.0",
    "gcsfs>=2026.2.0",
    "ijson>=3.4.0.post0",
    "pandas>=3.0.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-ai-slim>=1.58.0",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.2",
    "scikit-learn>=1.8.0",
    "umap-learn>=0.5.11",
    "sse-starlette>=3.2.0",
    "supervision>=0.27.0",
    "torch>=2.10.0",
    "transformers>=5.1.0",
    "uvicorn[standard]>=0.40.0",
    "pyvips>=3.1.1",
    "google-genai>=1.63.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
